+++ examples/dhcp.rs
@@ -3,34 +3,35 @@
 #![feature(c_variadic)]
 #![feature(const_mut_refs)]
 
-use ble_hci::{
-    ad_structure::{
-        create_advertising_data, AdStructure, BR_EDR_NOT_SUPPORTED, LE_GENERAL_DISCOVERABLE,
-    },
-    att::Uuid,
-    Ble, HciConnector,
-};
-use esp_wifi::ble::controller::BleConnector;
+#[cfg(feature = "esp32")]
+use esp32_hal as hal;
+#[cfg(feature = "esp32c3")]
+use esp32c3_hal as hal;
 
 use embedded_svc::wifi::{
     AccessPointInfo, ClientConfiguration, ClientConnectionStatus, ClientIpStatus, ClientStatus,
     Configuration, Status, Wifi,
 };
-use esp32c3_hal::clock::{ClockControl, CpuClock};
-use esp32c3_hal::system::SystemExt;
-use esp32c3_hal::systimer::SystemTimer;
-use esp32c3_hal::{pac::Peripherals, prelude::*, Rtc};
 use esp_backtrace as _;
 use esp_println::{print, println};
 use esp_wifi::initialize;
 use esp_wifi::wifi::utils::create_network_interface;
 use esp_wifi::wifi_interface::{timestamp, WifiError};
 use esp_wifi::{create_network_stack_storage, network_stack_storage};
-use riscv_rt::entry;
+use hal::clock::{ClockControl, CpuClock};
+use hal::{pac::Peripherals, prelude::*, Rtc};
 use smoltcp::{
     iface::SocketHandle,
     socket::{Socket, TcpSocket},
 };
+
+#[cfg(feature = "esp32c3")]
+use hal::system::SystemExt;
+
+#[cfg(feature = "esp32c3")]
+use riscv_rt::entry;
+#[cfg(feature = "esp32")]
+use xtensa_lx_rt::entry;
 
 extern crate alloc;
 
@@ -43,22 +44,41 @@ fn main() -> ! {
     esp_wifi::init_heap();
 
     let peripherals = Peripherals::take().unwrap();
+
+    #[cfg(not(feature = "esp32"))]
     let system = peripherals.SYSTEM.split();
+    #[cfg(feature = "esp32")]
+    let system = peripherals.DPORT.split();
+
+    #[cfg(feature = "esp32c3")]
     let clocks = ClockControl::configure(system.clock_control, CpuClock::Clock160MHz).freeze();
+    #[cfg(feature = "esp32")]
+    let clocks = ClockControl::configure(system.clock_control, CpuClock::Clock240MHz).freeze();
 
     let mut rtc = Rtc::new(peripherals.RTC_CNTL);
 
     // Disable watchdog timers
+    #[cfg(not(feature = "esp32"))]
     rtc.swd.disable();
+
     rtc.rwdt.disable();
 
     let mut storage = create_network_stack_storage!(3, 8, 1);
     let ethernet = create_network_interface(network_stack_storage!(storage));
     let mut wifi_interface = esp_wifi::wifi_interface::Wifi::new(ethernet);
 
-    let syst = SystemTimer::new(peripherals.SYSTIMER);
-
-    initialize(syst.alarm0, peripherals.RNG, &clocks).unwrap();
+    #[cfg(feature = "esp32c3")]
+    {
+        use hal::systimer::SystemTimer;
+        let syst = SystemTimer::new(peripherals.SYSTIMER);
+        initialize(syst.alarm0, peripherals.RNG, &clocks).unwrap();
+    }
+    #[cfg(feature = "esp32")]
+    {
+        use hal::timer::TimerGroup;
+        let timg1 = TimerGroup::new(peripherals.TIMG1, &clocks);
+        initialize(timg1.timer0, peripherals.RNG, &clocks).unwrap();
+    }
 
     println!("{:?}", wifi_interface.get_status());
 
@@ -91,25 +111,7 @@ fn main() -> ! {
     }
     println!("{:?}", wifi_interface.get_status());
 
-    let connector = BleConnector {};
-    let hci = HciConnector::new(connector, esp_wifi::current_millis);
-    let mut ble = Ble::new(&hci);
-
-    println!("{:?}", ble.init());
-    println!("{:?}", ble.cmd_set_le_advertising_parameters());
-    println!(
-        "{:?}",
-        ble.cmd_set_le_advertising_data(create_advertising_data(&[
-            AdStructure::Flags(LE_GENERAL_DISCOVERABLE | BR_EDR_NOT_SUPPORTED),
-            AdStructure::ServiceUuids16(&[Uuid::Uuid16(0x1809)]),
-            AdStructure::CompleteLocalName("ESP32C3 BLE"),
-        ]))
-    );
-    println!("{:?}", ble.cmd_set_le_advertise_enable(true));
-
-    println!("started advertising");
-
-    println!("Start busy loop on main - you can also scan for the BLE device");
+    println!("Start busy loop on main");
 
     let mut stage = 0;
     let mut idx = 0;
